using decimal for floating point accuracy

+++ compute.py
@@ -2,6 +2,7 @@
 # seunghoon.lee.010@gmail.com
 
 import sys
+from decimal import Decimal
 
 MAX_INPUT_NUM = 100
 
@@ -34,13 +35,13 @@ def compute(threshold, limit, nums):
         output = 0.0
         if num > threshold:
             output = num - threshold
-            room = limit-total
+            room = limit - total
             output = min(output, room)
             total += output
         
         ans.append(output)
 
-    ans.append(float(total))
+    ans.append(Decimal(total))
     return ans
 
 
@@ -55,12 +56,12 @@ if __name__ == '__main__':
             print("You need to provide at least two arguments in your command: python <filename> <threshold> <limit>. Every argument after will be ignored.")
             sys.exit()
 
-        threshold = float(args[1])
-        limit = float(args[2])
+        threshold = Decimal(args[1])
+        limit = Decimal(args[2])
 
         nums = []
         for line in sys.stdin:
-            nums.append(float(line))    
+            nums.append(Decimal(line))
 
         ans = compute(threshold, limit, nums)
